backend/requirements.txt
anyio==4.4.0
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.6.2
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.8
daphne==4.1.2
Django==4.2.6
django-cors-headers==4.3.0
django-countries==7.6.1
django-filter==23.3
django-money==3.5.3
django-soft-delete==1.0.13
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
gunicorn==21.2.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httptools==0.6.1
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.7
importlib-metadata==6.8.0
incremental==22.10.0
Markdown==3.5
numpy==2.0.0
openpyxl==3.1.5
ordered-set==4.1.0
packaging==24.1
pandas==2.2.2
priority==1.3.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
py-moneyed==3.0
requests==2.32.3
service-identity==24.1.0
setuptools==71.1.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.15
uvicorn==0.30.3
watchfiles==0.22.0
websockets==12.0
whitenoise==6.6.0
zipp==3.17.0
zope.interface==6.4.post2
drf-spectacular==0.27.2
pydantic==2.8.2
pydantic-core==2.20.1
email-validator==2.3.0
dnspython==2.7.0
black==24.8.0
ruff==0.6.2